+++ packages/common-cosmos/src/project/versioned/v1_0_0/model.ts
@@ -13,7 +13,6 @@ import {
 } from '@subql/common';
 import {NodeSpec, ParentProject, QuerySpec, RunnerSpecs} from '@subql/types-core';
 import {
-  CustomModule,
   SubqlCosmosCustomDatasource,
   SubqlCosmosCustomHandler,
   SubqlCosmosMapping,
@@ -22,18 +21,10 @@ import {
   RuntimeDatasourceTemplate,
   CustomDatasourceTemplate,
   CosmosProjectManifestV1_0_0,
+  CosmosChainTypes,
 } from '@subql/types-cosmos';
 import {Transform, TransformFnParams, Type} from 'class-transformer';
-import {
-  Equals,
-  IsObject,
-  IsString,
-  ValidateNested,
-  validateSync,
-  IsOptional,
-  IsArray,
-  IsNotEmpty,
-} from 'class-validator';
+import {Equals, IsObject, IsString, ValidateNested, IsOptional, IsArray, IsNotEmpty} from 'class-validator';
 import {CosmosCustomDataSourceBase, CosmosCustomModuleImpl, CosmosRuntimeDataSourceBase} from '../../models';
 
 const COSMOS_NODE_NAME = `@subql/node-cosmos`;
@@ -72,13 +63,17 @@ export class CosmosProjectNetworkDeployment {
   @IsOptional()
   @IsArray()
   bypassBlocks?: (number | string)[];
+  @IsOptional()
+  @Type(() => CosmosCustomModuleImpl)
+  @ValidateNested({each: true})
+  chainTypes?: Map<string, CosmosCustomModuleImpl>;
 }
 
-export class CosmosProjectNetwork extends CommonProjectNetworkV1_0_0<FileType> {
+export class CosmosProjectNetwork extends CommonProjectNetworkV1_0_0<CosmosChainTypes> {
   @Type(() => CosmosCustomModuleImpl)
   @IsOptional()
   @ValidateNested({each: true})
-  chainTypes?: Map<string, CustomModule>;
+  chainTypes?: Map<string, CosmosCustomModuleImpl>;
 }
 
 export class RuntimeDatasourceTemplateImpl extends CosmosRuntimeDataSourceImpl implements RuntimeDatasourceTemplate {